+++ src/instruction.rs
@@ -1,61 +1,6 @@
 #![allow(dead_code)]
 
-#[derive(Debug, PartialEq, Clone)]
-pub enum Opcode {
-    HLT,
-    IGL,
-    LOAD,
-    ADD,
-    SUB,
-    MUL,
-    DIV,
-    /// Equal
-    EQ,
-    /// Not equal
-    NEQ,
-    /// Greater then
-    GT,
-    /// Less then
-    LT,
-    /// Greater then or equal to
-    GTE,
-    /// less then or equal
-    LTE,
-    /// jump if equal
-    JMPE,
-    /// Jump
-    JMP,
-    /// Jump forward
-    JMPF,
-    /// Jump backward
-    JMPB,
-    NOP,
-}
-
-impl From<u8> for Opcode {
-    fn from(vm: u8) -> Self {
-        match vm {
-            0 => Opcode::LOAD,
-            1 => Opcode::ADD,
-            2 => Opcode::SUB,
-            3 => Opcode::MUL,
-            4 => Opcode::DIV,
-            5 => Opcode::HLT,
-            6 => Opcode::JMP,
-            7 => Opcode::JMPF,
-            8 => Opcode::JMPB,
-            9 => Opcode::EQ,
-            10 => Opcode::NEQ,
-            11 => Opcode::GTE,
-            12 => Opcode::GT,
-            13 => Opcode::LTE,
-            14 => Opcode::LT,
-            15 => Opcode::JMPE,
-            16 => Opcode::NOP,
-            _ => Opcode::IGL,
-        }
-    }
-}
+use crate::assembler::Opcode;
 
 #[derive(Debug, PartialEq)]
 pub struct Instruction {
@@ -71,6 +16,7 @@ impl Instruction {
 #[cfg(test)]
 mod instruction_tests {
     use crate::instruction::*;
+    use crate::assembler::Opcode;
 
     #[test]
     fn test_crate_hlt() {
